+++ offlineimap/folder/LocalStatusSQLite.py
@@ -46,12 +46,8 @@ class LocalStatusSQLiteFolder(LocalStatusFolder):
     #current version of our db format
     cur_version = 1
 
-    def __init__(self, root, name, repository, accountname, config):
-        super(LocalStatusSQLiteFolder, self).__init__(root, name, 
-                                                      repository, 
-                                                      accountname,
-                                                      config)       
-
+    def __init__(self, name, repository):
+        super(LocalStatusSQLiteFolder, self).__init__(name, repository)       
         # dblock protects against concurrent writes in same connection
         self._dblock = Lock()
         #Try to establish connection, no need for threadsafety in __init__