+++ offlineimap/init.py
@@ -24,19 +24,16 @@ import signal
 import socket
 import logging
 from optparse import OptionParser
+try:
+    import fcntl
+except ImportError:
+    pass #it's OK
 import offlineimap
 from offlineimap import accounts, threadutil, syncmaster
+from offlineimap.error import OfflineImapError
 from offlineimap.ui import UI_LIST, setglobalui, getglobalui
 from offlineimap.CustomConfig import CustomConfigParser
 
-
-try:
-    import fcntl
-    hasfcntl = 1
-except:
-    hasfcntl = 0
-
-lockfd = None
 
 class OfflineImap:
     """The main class that encapsulates the high level use of OfflineImap.
@@ -46,17 +43,6 @@ class OfflineImap:
       oi = OfflineImap()
       oi.run()
     """
-    def lock(self, config, ui):
-        global lockfd, hasfcntl
-        if not hasfcntl:
-            return
-        lockfd = open(config.getmetadatadir() + "/lock", "w")
-        try:
-            fcntl.flock(lockfd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-        except IOError:
-            ui.locked()
-            ui.terminate(1)
-    
     def run(self):
         """Parse the commandline and invoke everything"""
 
@@ -253,7 +239,6 @@ class OfflineImap:
                     config.set(section, "folderfilter", folderfilter)
                     config.set(section, "folderincludes", folderincludes)
 
-        self.lock(config, ui)
         self.config = config
     
         def sigterm_handler(signum, frame):
@@ -330,6 +315,18 @@ class OfflineImap:
             #various initializations that need to be performed:
             offlineimap.mbnames.init(config, syncaccounts)
 
+            #TODO: keep legacy lock for a few versions, then remove.
+            self._legacy_lock = open(self.config.getmetadatadir() + "/lock",
+                                       'w')
+            try:
+                fcntl.lockf(self._legacy_lock, fcntl.LOCK_EX|fcntl.LOCK_NB)
+            except NameError:
+                #fcntl not available (Windows), disable file locking... :(
+                pass
+            except IOError:
+                raise OfflineImapError("Could not take global lock.",
+                                       OfflineImapError.ERROR.REPO)
+
             if options.singlethreading:
                 #singlethreaded
                 self.sync_singlethreaded(syncaccounts, config)
@@ -349,8 +346,9 @@ class OfflineImap:
             return
         except (SystemExit):
             raise
-        except:
-            ui.mainException()
+        except Exception, e:
+            ui.error(e)
+            ui.terminate()
 
     def sync_singlethreaded(self, accs, config):
         """Executed if we do not want a separate syncmaster thread